requirements.txt
pymongo==4.15.3
python-dotenv==1.1.1
kafka-python==2.2.15
apache-beam==2.68.0
psycopg2-binary==2.9.11
sqlalchemy==2.0.44